feat: Implement PPH management with CRUD operations and validation

+++ database/migrations/2025_11_04_143026_create_payment_vouchers_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->foreignId("supplier_id")->constrained("suppliers")->cascadeOnDelete();
             $table->foreignId("supplier_account_id")->constrained("supplier_accounts")->cascadeOnDelete();
             $table->morphs("processable");
-            $table->unsignedInteger("pv_amount");
+            $table->unsignedBigInteger("pv_amount");
             $table->string("status")->default("NEW");
             $table->date("paid_date")->nullable();
             $table->foreignId("trx_dtl_id")->constrained("type_trxes")->cascadeOnDelete();
